refactored search message sidebar
